requirements.txt
google-cloud-spanner
google-cloud-aiplatform
streamlit
faker
networkx
pyvis
pandas
numpy
google-auth
